refactor tests to use default init18n
